Split out the modifiers into their own files
