+++ src/definition/accessors/index.ts
@@ -14,26 +14,25 @@ import {
     IHttpResponse,
     RequestMethod,
 } from './IHttp';
+import { ILivechatCreator } from './ILivechatCreator';
+import { ILivechatMessageBuilder } from './ILivechatMessageBuilder';
 import { ILivechatRead } from './ILivechatRead';
+import { ILivechatUpdater } from './ILivechatUpdater';
 import { ILogEntry, LogMessageSeverity } from './ILogEntry';
 import { ILogger } from './ILogger';
+import { IMessageBuilder } from './IMessageBuilder';
+import { IMessageExtender } from './IMessageExtender';
 import { IMessageRead } from './IMessageRead';
-import {
-    ILivechatMessageBuilder,
-    ILivechatUpdater,
-    IMessageBuilder,
-    IMessageExtender,
-    IModify,
-    IModifyCreator,
-    IModifyExtender,
-    IModifyUpdater,
-    INotifier,
-    IRoomBuilder,
-    IRoomExtender,
-} from './IModify';
+import { IModify } from './IModify';
+import { IModifyCreator } from './IModifyCreator';
+import { IModifyExtender } from './IModifyExtender';
+import { IModifyUpdater } from './IModifyUpdater';
+import { INotifier } from './INotifier';
 import { IPersistence } from './IPersistence';
 import { IPersistenceRead } from './IPersistenceRead';
 import { IRead } from './IRead';
+import { IRoomBuilder } from './IRoomBuilder';
+import { IRoomExtender } from './IRoomExtender';
 import { IRoomRead } from './IRoomRead';
 import { IServerSettingRead } from './IServerSettingRead';
 import { IServerSettingsModify } from './IServerSettingsModify';
@@ -57,6 +56,7 @@ export {
     IHttpPreResponseHandler,
     IHttpRequest,
     IHttpResponse,
+    ILivechatCreator,
     ILivechatMessageBuilder,
     ILivechatRead,
     ILivechatUpdater,
